client supports TLS
